requirements.txt
pypdf
python-docx
beautifulsoup4
sentence-transformers
numpy
